Refactor card layout and add split-hand support
- Remove deprecated CenteredCardStackView.swift
- Introduce FanCardsView.swift for improved fan-style card stacking
- Update DevMenu: add “Deal Split Cards” button and .devDealSplitCards notification
- Implement devDealSplitCards() in NewGameView to simulate split-hand dealing
- Expose isSplitView state and add NewGameViewModel (@Observable) for future view logic
- Tweak FlipCardView rotation comment for clarity
